Add files via upload
Tic Tac Toe is a simple game that is played on a 3×3 grid. The game is played by two players, “X” and “O”, who take turns marking the spaces in the grid. The player who succeeds in placing three of their marks in a horizontal, vertical, or diagonal row wins the game.